Gross Salary Finder
This is a project in C language in which user can find his/her gross salary if the dear allowance of the user is 40% and the rent allowance of the user is 20%. User can also change the main file to add or change the features of the program. This my 1st program!